requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
pydantic>=2.0.0
markdown>=3.5.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
Pygments>=2.17.0
